Added : Ability to delete employee from json server by name

+++ src/main/java/com/capgemini/payrollservice/EmployeePayrollService.java
@@ -124,10 +124,15 @@ public class EmployeePayrollService {
 				employeePayrollDBService.addEmployeeToPayroll(company_id, name, gender, salary, start, departmentList));
 	}
 
-	public void deleteEmployeeInDatabase(String name) throws DataBaseSQLException {
-		boolean result = this.employeePayrollDBService.deleteEmployee(name);
-		System.out.println(result);
-		if (result) {
+	public void deleteEmployeeInDatabase(String name, IOService ioservice) throws DataBaseSQLException {
+		if (ioservice.equals(IOService.DB_IO)) {
+			boolean result = this.employeePayrollDBService.deleteEmployee(name);
+			System.out.println(result);
+			if (result) {
+				EmployeePayrollData emp = getEmployee(employeePayrollList, name);
+				employeePayrollList.remove(emp);
+			}
+		} else {
 			EmployeePayrollData emp = getEmployee(employeePayrollList, name);
 			employeePayrollList.remove(emp);
 		}
